qca-nss-clients: Connection managers for higher level transports
Currently the following is confirmed to be able to be compiled:

qca-nss-drv-dtlsmgr
qca-nss-drv-eogremgr
qca-nss-drv-gre
qca-nss-drv-ipsecmgr
qca-nss-drv-pppoe
qca-nss-drv-qdisc
qca-nss-drv-tun6rd
qca-nss-drv-tunipip6

Qdisc appears to work, as least for nsstbl, nsscodel & nssfq_codel.  The
rest of he qdisc algo should work as well.

The rest of the compiled modules have to be further tested.